# backend/models/User.ts
import mongoose, { Document, Schema } from 'mongoose';
import bcrypt from 'bcryptjs';

// Force clean up pour éviter les erreurs de validation avec l'ancien modèle
if (mongoose.models.User) {
  delete mongoose.models.User;
}

export interface IUser extends Document {
  email: string;
  password: string;
  firstName: string;
  lastName: string;
  address: string;
  phone: string;
  birthDate: Date;
  city: string;
  role: 'user' | 'admin';
  createdAt: Date;
  updatedAt: Date;
  comparePassword(candidatePassword: string): Promise<boolean>;
  isModified(path: string): boolean;
}

const UserSchema = new Schema<IUser>(
  {
    email: {
      type: String,
      required: [true, 'Email est requis'],
      unique: true,
      trim: true,
      lowercase: true,
      match: [/^\S+@\S+\.\S+$/, 'Format email invalide'],
    },
    password: {
      type: String,
      required: [true, 'Mot de passe est requis'],
      minlength: [6, 'Le mot de passe doit contenir au moins 6 caractères'],
    },
    firstName: {
      type: String,
      required: [true, 'Prénom est requis'],
      trim: true,
    },
    lastName: {
      type: String,
      required: [true, 'Nom est requis'],
      trim: true,
    },
    address: {
      type: String,
      required: [true, 'Adresse est requise'],
      trim: true,
    },
    phone: {
      type: String,
      required: [true, 'Numéro de téléphone est requis'],
      trim: true,
    },
    birthDate: {
      type: Date,
      required: [true, 'Date de naissance est requise'],
    },
    city: {
      type: String,
      required: [true, 'Ville est requise'],
      trim: true,
    },
    role: {
      type: String,
      enum: ['user', 'admin'],
      default: 'user',
    },
  },
  {
    timestamps: true,
  }
);

// Hash le mot de passe avant de sauvegarder
UserSchema.pre('save', async function(this: IUser, next: mongoose.CallbackWithoutResultAndOptionalError) {
  if (!this.isModified('password')) return next();
  
  try {
    const salt = await bcrypt.genSalt(10);
    this.password = await bcrypt.hash(this.password, salt);
    next();
  } catch (error: any) {
    next(error);
  }
});

// Méthode pour comparer les mots de passe
UserSchema.methods.comparePassword = async function(this: IUser, candidatePassword: string): Promise<boolean> {
  return bcrypt.compare(candidatePassword, this.password);
};

// Créer un nouveau modèle (après avoir supprimé l'ancien si présent)
const User = mongoose.model<IUser>('User', UserSchema);

export default User; 